v0.7

+++ src/Core/App.hpp
@@ -3,6 +3,7 @@
 #include "SDL.h"
 #include "../Resources/Resources.hpp"
 #include <iostream>
+#include "Assert.h"
 #include "../Renderer/Atlas.hpp"
 
 class Scene;
@@ -59,5 +60,5 @@ private:
 	vec3f m_window_color = vec3f(0, 0, 0);
 
 	//System
-	float m_fps = 0;
+	int m_fps = 0;
 };
